+++ .github/workflows/create-release.js
@@ -4,7 +4,7 @@ module.exports = main;
 
 const CHANGELOG_FILE = 'CHANGELOG.md';
 const MAX_BODY_LENGTH = 125000;
-const LOCAL = ['zh-CN', {timeZone: 'CST'}, ' (北京时间)'];
+const LOCAL = ['zh-CN', {timeZone: 'CST'}, ' (Beijing UTC+8)'];
 
 
 async function main({github, context, sha}) {